+++ src/kafkit/__init__.py
@@ -1,10 +1,8 @@
-"""Kafkit helps you write Kafka producers and consumers in Python with asyncio.
-"""
+"""Kafkit helps you write Kafka producers and consumers in with asyncio."""
 
 __all__ = ["__version__", "version_info"]
 
 from importlib.metadata import PackageNotFoundError, version
-from typing import List
 
 __version__: str
 """The version string of Kafkit (PEP 440 / SemVer compatible)."""
@@ -15,7 +13,7 @@ except PackageNotFoundError:
     # package is not installed
     __version__ = "0.0.0"
 
-version_info: List[str] = __version__.split(".")
+version_info = __version__.split(".")
 """The decomposed version, split across "``.``."
 
 Use this for version comparison.