+++ src/kafkit/registry/sansio.py
@@ -9,11 +9,13 @@ See licenses/gidgethub.txt for license info.
 from __future__ import annotations
 
 import abc
+import contextlib
 import copy
 import json
 import logging
+from collections.abc import Mapping
 from enum import Enum
-from typing import Any, Dict, Mapping, Optional, Tuple, Union, overload
+from typing import Any, ClassVar, cast, overload
 
 import fastavro
 
@@ -26,18 +28,18 @@ from kafkit.registry.errors import (
 )
 
 __all__ = [
-    "make_headers",
-    "decipher_response",
-    "decode_body",
-    "RegistryApi",
+    "CompatibilityType",
     "MockRegistryApi",
+    "RegistryApi",
     "SchemaCache",
     "SubjectCache",
-    "CompatibilityType",
+    "decipher_response",
+    "decode_body",
+    "make_headers",
 ]
 
 
-def make_headers() -> Dict[str, str]:
+def make_headers() -> dict[str, str]:
     """Make HTTP headers for the Confluent Schema Registry.
 
     Returns
@@ -46,8 +48,7 @@ def make_headers() -> Dict[str, str]:
         A dictionary of HTTP headers for a Confluent Schema Registry request.
         All keys are normalized to lowercase for consistency.
     """
-    headers = {"accept": "application/vnd.schemaregistry.v1+json"}
-    return headers
+    return {"accept": "application/vnd.schemaregistry.v1+json"}
 
 
 def decipher_response(
@@ -72,17 +73,16 @@ def decipher_response(
             raise RegistryBrokenError(
                 status_code=status_code, error_code=error_code, message=message
             )
-        elif status_code >= 400:
+        if status_code >= 400:
             raise RegistryBadRequestError(
                 status_code=status_code, error_code=error_code, message=message
             )
-        elif status_code >= 300:
+        if status_code >= 300:
             raise RegistryRedirectionError(status_code=status_code)
-        else:
-            raise RegistryHttpError(status_code=status_code)
+        raise RegistryHttpError(status_code=status_code)
 
 
-def decode_body(content_type: Optional[str], body: bytes) -> Any:
+def decode_body(content_type: str | None, body: bytes) -> Any:
     """Decode an HTTP body based on the specified content type.
 
     Parameters
@@ -142,7 +142,7 @@ class RegistryApi(metaclass=abc.ABCMeta):
     @abc.abstractmethod
     async def _request(
         self, method: str, url: str, headers: Mapping[str, str], body: bytes
-    ) -> Tuple[int, Mapping[str, str], bytes]:
+    ) -> tuple[int, Mapping[str, str], bytes]:
         """Make an HTTP request.
 
         Parameters
@@ -171,19 +171,18 @@ class RegistryApi(metaclass=abc.ABCMeta):
         else:
             charset = "utf-8"
             body = json.dumps(data).encode(charset)
-            request_headers[
-                "content-type"
-            ] = f"application/json; charset={charset}"
+            request_headers["content-type"] = (
+                f"application/json; charset={charset}"
+            )
             request_headers["content-length"] = str(len(body))
 
         response = await self._request(
             method, expanded_url, request_headers, body
         )
-        response_data = decipher_response(*response)
-        return response_data
+        return decipher_response(*response)
 
     async def get(
-        self, url: str, url_vars: Optional[Mapping[str, str]] = None
+        self, url: str, url_vars: Mapping[str, str] | None = None
     ) -> Any:
         """Send an HTTP GET request.
 
@@ -216,13 +215,12 @@ class RegistryApi(metaclass=abc.ABCMeta):
         """
         if url_vars is None:
             url_vars = {}
-        data = await self._make_request("GET", url, url_vars, b"")
-        return data
+        return await self._make_request("GET", url, url_vars, b"")
 
     async def post(
         self,
         url: str,
-        url_vars: Optional[Mapping[str, str]] = None,
+        url_vars: Mapping[str, str] | None = None,
         *,
         data: Any,
     ) -> Any:
@@ -259,13 +257,12 @@ class RegistryApi(metaclass=abc.ABCMeta):
         """
         if url_vars is None:
             url_vars = {}
-        data = await self._make_request("POST", url, url_vars, data)
-        return data
+        return await self._make_request("POST", url, url_vars, data)
 
     async def patch(
         self,
         url: str,
-        url_vars: Optional[Mapping[Any, Any]] = None,
+        url_vars: Mapping[Any, Any] | None = None,
         *,
         data: Any,
     ) -> Any:
@@ -302,13 +299,12 @@ class RegistryApi(metaclass=abc.ABCMeta):
         """
         if url_vars is None:
             url_vars = {}
-        data = await self._make_request("PATCH", url, url_vars, data)
-        return data
+        return await self._make_request("PATCH", url, url_vars, data)
 
     async def put(
         self,
         url: str,
-        url_vars: Optional[Mapping[str, str]] = None,
+        url_vars: Mapping[str, str] | None = None,
         data: Any = b"",
     ) -> Any:
         """Send an HTTP PUT request.
@@ -344,13 +340,12 @@ class RegistryApi(metaclass=abc.ABCMeta):
         """
         if url_vars is None:
             url_vars = {}
-        data = await self._make_request("PUT", url, url_vars, data)
-        return data
+        return await self._make_request("PUT", url, url_vars, data)
 
     async def delete(
         self,
         url: str,
-        url_vars: Optional[Mapping[str, str]] = None,
+        url_vars: Mapping[str, str] | None = None,
         *,
         data: Any = b"",
     ) -> Any:
@@ -385,8 +380,7 @@ class RegistryApi(metaclass=abc.ABCMeta):
         """
         if url_vars is None:
             url_vars = {}
-        data = await self._make_request("DELETE", url, url_vars, data)
-        return data
+        return await self._make_request("DELETE", url, url_vars, data)
 
     @staticmethod
     def _prep_schema(schema: Mapping[str, Any]) -> str:
@@ -402,7 +396,7 @@ class RegistryApi(metaclass=abc.ABCMeta):
         return json.dumps(schema, sort_keys=True)
 
     async def register_schema(
-        self, schema: Mapping[str, Any], subject: Optional[str] = None
+        self, schema: Mapping[str, Any], subject: str | None = None
     ) -> int:
         """Register a schema or get the ID of an existing schema.
 
@@ -411,8 +405,7 @@ class RegistryApi(metaclass=abc.ABCMeta):
         Parameters
         ----------
         schema : `dict`
-            An `Avro schema <http://avro.apache.org/docs/current/spec.html>`__
-            as a Python dictionary.
+            An Avro schema as a Python dictionary.
         subject : `str`, optional
             The subject to register the schema under. If not provided, the
             fully-qualified name of the schema is adopted as the subject name.
@@ -430,36 +423,36 @@ class RegistryApi(metaclass=abc.ABCMeta):
         """
         # Parsing the schema also produces a fully-qualified name, which is
         # useful for getting a subject name.
-        schema = fastavro.parse_schema(schema)
+        parsed_schema = fastavro.parse_schema(dict(schema))
+        parsed_schema = cast("dict[str, Any]", parsed_schema)
 
         # look in cache first
         try:
-            schema_id = self.schema_cache[schema]
-            return schema_id
+            return self.schema_cache[parsed_schema]
         except KeyError:
             pass
 
         if subject is None:
             try:
-                subject = schema["name"]
-            except (KeyError, TypeError):
+                subject = parsed_schema["name"]
+            except (KeyError, TypeError) as e:
                 raise RuntimeError(
                     "Cannot get a subject name from a 'name' "
-                    f"key in the schema: {schema!r}"
-                )
+                    f"key in the schema: {parsed_schema!r}"
+                ) from e
 
         result = await self.post(
             "/subjects{/subject}/versions",
             url_vars={"subject": subject},
-            data={"schema": self._prep_schema(schema)},
+            data={"schema": self._prep_schema(parsed_schema)},
         )
 
         # add to cache
-        self.schema_cache.insert(schema, result["id"])
+        self.schema_cache.insert(parsed_schema, result["id"])
 
         return result["id"]
 
-    async def get_schema_by_id(self, schema_id: int) -> Dict[str, Any]:
+    async def get_schema_by_id(self, schema_id: int) -> dict[str, Any]:
         """Get a schema from the registry given its ID.
 
         Wraps ``GET /schemas/ids/{int: id}``.
@@ -487,8 +480,7 @@ class RegistryApi(metaclass=abc.ABCMeta):
         """
         # Look in the cache first
         try:
-            schema = self.schema_cache[schema_id]
-            return schema
+            return self.schema_cache[schema_id]
         except KeyError:
             pass
 
@@ -496,6 +488,7 @@ class RegistryApi(metaclass=abc.ABCMeta):
             "/schemas/ids{/schema_id}", url_vars={"schema_id": str(schema_id)}
         )
         schema = fastavro.parse_schema(json.loads(result["schema"]))
+        schema = cast("dict[str, Any]", schema)
 
         # Add schema to cache
         self.schema_cache.insert(schema, schema_id)
@@ -503,8 +496,8 @@ class RegistryApi(metaclass=abc.ABCMeta):
         return schema
 
     async def get_schema_by_subject(
-        self, subject: str, version: Union[str, int] = "latest"
-    ) -> Dict[str, Any]:
+        self, subject: str, version: str | int = "latest"
+    ) -> dict[str, Any]:
         """Get a schema for a subject in the registry.
 
         Wraps ``GET /subjects/(string: subject)/versions/(versionId: version)``
@@ -559,17 +552,16 @@ class RegistryApi(metaclass=abc.ABCMeta):
         )
 
         schema = fastavro.parse_schema(json.loads(result["schema"]))
+        schema = cast("dict[str, Any]", schema)
 
-        try:
+        with contextlib.suppress(TypeError):
+            # Can't cache versions like "latest"
             self.subject_cache.insert(
                 result["subject"],
                 result["version"],
                 schema_id=result["id"],
                 schema=schema,
             )
-        except TypeError:
-            # Can't cache versions like "latest"
-            pass
 
         return {
             "id": result["id"],
@@ -584,7 +576,7 @@ class MockRegistryApi(RegistryApi):
     network operations and provides attributes for introspection.
     """
 
-    DEFAULT_HEADERS = {
+    DEFAULT_HEADERS: ClassVar = {
         "content-type": "application/vnd.schemaregistry.v1+json"
     }
 
@@ -592,7 +584,7 @@ class MockRegistryApi(RegistryApi):
         self,
         url: str = "http://registry:8081",
         status_code: int = 200,
-        headers: Optional[Mapping[str, str]] = None,
+        headers: Mapping[str, str] | None = None,
         body: Any = b"",
     ) -> None:
         super().__init__(url=url)
@@ -622,8 +614,8 @@ class SchemaCache:
     """
 
     def __init__(self) -> None:
-        self._id_to_schema: Dict[int, str] = {}
-        self._schema_to_id: Dict[str, int] = {}
+        self._id_to_schema: dict[int, str] = {}
+        self._schema_to_id: dict[str, int] = {}
 
     def insert(self, schema: Mapping[str, Any], schema_id: int) -> None:
         """Insert a schema into the cache.
@@ -643,16 +635,14 @@ class SchemaCache:
         self._schema_to_id[serialized_schema] = schema_id
 
     @overload
-    def __getitem__(self, key: int) -> Dict[str, Any]:
-        ...
+    def __getitem__(self, key: int) -> dict[str, Any]: ...
 
-    @overload  # noqa: F811 remove for pyflakes 2.2.x
-    def __getitem__(self, key: Mapping[str, Any]) -> int:  # noqa: F811
-        ...
+    @overload
+    def __getitem__(self, key: Mapping[str, Any]) -> int: ...
 
-    def __getitem__(  # noqa: F811 remove for pyflakes 2.2.x
-        self, key: Union[Mapping[str, Any], int]
-    ) -> Union[Dict[str, Any], int]:
+    def __getitem__(
+        self, key: Mapping[str, Any] | int
+    ) -> dict[str, Any] | int:
         if isinstance(key, int):
             return json.loads(self._id_to_schema[key])
         else:
@@ -661,15 +651,15 @@ class SchemaCache:
             schema = copy.deepcopy(key)
             try:
                 serialized_schema = SchemaCache._serialize_schema(schema)
-            except Exception:
+            except Exception as e:
                 # If the schema couldn't be parsed, its not going to be a
                 # valid key anyhow.
                 raise KeyError(
                     f"Key or schema not in the SchemaCache: {key!r}"
-                )
+                ) from e
             return self._schema_to_id[serialized_schema]
 
-    def __contains__(self, key: Union[int, Mapping[str, Any]]) -> bool:
+    def __contains__(self, key: int | Mapping[str, Any]) -> bool:
         try:
             self[key]
         except KeyError:
@@ -679,8 +669,9 @@ class SchemaCache:
     @staticmethod
     def _serialize_schema(schema: Mapping[str, Any]) -> str:
         """Predictably serialize the schema so that it's hashable."""
-        schema = fastavro.parse_schema(schema)
-        return json.dumps(schema, sort_keys=True)
+        parsed_schema = fastavro.parse_schema(dict(schema))
+        parsed_schema = cast("dict[str, Any]", parsed_schema)
+        return json.dumps(parsed_schema, sort_keys=True)
 
 
 class SubjectCache:
@@ -705,7 +696,7 @@ class SubjectCache:
     def __init__(self, schema_cache: SchemaCache) -> None:
         self.schema_cache = schema_cache
 
-        self._subject_to_id: Dict[Tuple[str, int], int] = {}
+        self._subject_to_id: dict[tuple[str, int], int] = {}
 
     def get_id(self, subject: str, version: int) -> int:
         """Get the schema ID of a subject version.
@@ -737,7 +728,7 @@ class SubjectCache:
         except KeyError as e:
             raise ValueError from e
 
-    def get_schema(self, subject: str, version: int) -> Dict[str, Any]:
+    def get_schema(self, subject: str, version: int) -> dict[str, Any]:
         """Get the schema of a subject version.
 
         Parameters
@@ -764,12 +755,11 @@ class SubjectCache:
         get
         """
         try:
-            schema = self.schema_cache[self.get_id(subject, version)]
-            return schema
+            return self.schema_cache[self.get_id(subject, version)]
         except KeyError as e:
             raise ValueError from e
 
-    def get(self, subject: str, version: Union[int, str]) -> Dict[str, Any]:
+    def get(self, subject: str, version: int | str) -> dict[str, Any]:
         """Get the full set of schema and ID information for a subject version.
 
         Parameters
@@ -807,7 +797,9 @@ class SubjectCache:
         get_schema
         """
         if not isinstance(version, int):
-            raise ValueError("version must be an int, got {}".format(version))
+            raise ValueError(  # noqa: TRY004
+                f"version must be an int, got {version}"
+            )
         try:
             schema_id = self.get_id(subject, version)
             schema = self.schema_cache[schema_id]
@@ -827,8 +819,8 @@ class SubjectCache:
         self,
         subject: str,
         version: int,
-        schema_id: Optional[int] = None,
-        schema: Optional[Mapping[str, Any]] = None,
+        schema_id: int | None = None,
+        schema: Mapping[str, Any] | None = None,
     ) -> None:
         """Insert a subject version into the cache.
 
@@ -896,7 +888,7 @@ class SubjectCache:
                 "Provide either a schema_id or schema argument (or both)."
             )
 
-    def __contains__(self, key: Tuple[str, int]) -> bool:
+    def __contains__(self, key: tuple[str, int]) -> bool:
         return key in self._subject_to_id